refactor(global): layout refactor and separation of styles
