+++ src/App.js
@@ -1,10 +1,12 @@
 import React from 'react';
 
+import GlobalStyles from './styles/global';
 import Routes from './routes';
 
 function App() {
   return (
     <>
+      <GlobalStyles />
       <Routes />
     </>
   );
