+++ src/components/Layout/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { Wrapper } from './styles';
+
+const Layout = ({ children }) => {
+  return (
+    <Wrapper>
+      <main>{children}</main>
+    </Wrapper>
+  );
+};
+
+export default Layout;
