+++ src/components/Layout/styles.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  width: 100%;
+  background-color: #000000;
+  color: white;
+  min-height: 100vh;
+
+  font-size: 16px;
+`;
