--- src/components/fonts.css
@@ -1,2 +1,0 @@
-/* Fonts */
-@import url("https://fonts.googleapis.com/css?family=Karla|Roboto&display=swap");
