+++ src/pages/Home/styles.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  width: 100%;
+  height: 100vh;
+
+  background-image: url(bgImage);
+  background-position: 200px;
+  background-repeat: no-repeat;
+  background-size: cover;
+`;
+
+export const Wrapper = styled.div`
+  width: 100%;
+  max-width: 1200px;
+  margin: 0 auto;
+`;
