+++ src/routes.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import IndexPage from './pages/IndexPage';
-import BindsPage from './pages/BindsPage';
-import BindsCFGPage from './pages/BindsCFGPage';
+import Home from './pages/Home';
+import Binds from './pages/Binds';
+import BindsCFG from './pages/BindsCFG';
 
 export default function Routes() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route exact path="/" component={IndexPage} />
-        <Route path="/binds" component={BindsPage} />
-        <Route path="/binds-cfg" component={BindsCFGPage} />
+        <Route exact path="/" component={Home} />
+        <Route path="/binds" component={Binds} />
+        <Route path="/binds-cfg" component={BindsCFG} />
       </Switch>
     </BrowserRouter>
   );
